requirements.txt
langchain-ollama
pillow
ipython